// src/app/model/transaction.ts
export interface Transaction {
  id: string;
  date: Date;
  ticker: string;
  rationaleForEntry: string;
  marketCondition: string;
  patternNumber: number;
  volumeAvailability: string;
  wayOut: string;
  evaluationOfEntrance: string;
  maxPossibleProfitFactor: string;
  transactionResult: number;
  deposit: number;
  mistakes: string;
}